feature: A media file sub-directory can now be copied in place of a media file or in conjunction with a media file.

+++ lib/mymedia-card.rb
@@ -13,8 +13,7 @@ class MyMediaCard < MyMediaKvx
                                                      config: nil, ext: 'txt')
     
     super(media_type: media_type, public_type: public_type, 
-                                                    config: config, ext: ext)
-    
+                                                    config: config, ext: ext)    
     @prefix = 'meta'
     @txt_filepath = File.join(@media_src, \
                               Time.now.strftime('meta%d%m%yT%H%M') + '.txt' )
@@ -29,10 +28,12 @@ class MyMediaCard < MyMediaKvx
     raw_s = dir.select_by_ext(@ext).sort_by_lastmodified.last[:name]
     
     s,ext = raw_s.split(/(?=\.\w+$)/)
+    ext = '' if ext == '.dir'
     raw_name, raw_tags = s.split(/--/,2)
     tags = raw_tags ? raw_tags.split(/--/) : []
     desc = raw_name.gsub(/-/,' ')
     desc.capitalize! unless desc[0] == desc[0].upcase
+    
     filename = raw_name.downcase + ext
 
     static_path = "%s/%s/%s" % [@public_type, \
@@ -43,12 +44,11 @@ class MyMediaCard < MyMediaKvx
     summary = {title: desc, tags: tags.join(' ')}
     body = {file: raw_static_destination}    
 
-    kvx = Kvx.new({summary: summary, body: body}, attributes: {type: @media_type})
-    
+    kvx = Kvx.new({summary: summary, body: body}, \
+                                             attributes: {type: @media_type})
     Dir.chdir @media_src
-    
     File.write @txt_filepath, kvx.to_s
-    
+
     FileUtils.mv raw_s, filename
   end
   
@@ -58,11 +58,26 @@ class MyMediaCard < MyMediaKvx
   def copy_edit(src_path, destination, raw='')
 
     kvx, raw_msg = super(src_path, destination)
-    
+
     # copy the media file to the destination
     destination = kvx.body[:file]
-    file = destination[/[^\/]+$/]
-    FileUtils.cp File.join(@media_src, file), File.join(@home, destination)
+    file = File.basename(destination)
+    
+    subdir = file[/.*(?=\.\w+$)|.*/]
+    
+    if File.exists? File.join(@media_src, subdir) then
+      
+      FileUtils.cp_r File.join(@media_src, subdir), File.join(@home, File.dirname(destination))
+      
+      if File.basename(destination) != subdir then
+        FileUtils.cp File.join(@media_src, file), File.join(@home, destination)
+      end
+      
+    else
+      
+      FileUtils.cp File.join(@media_src, file), File.join(@home, destination)
+      
+    end
     
     [kvx, raw_msg]
   end
